feat: reorganize structure, create utils module.
