+++ src/readii/io/utils/__init__.py
@@ -1,0 +1,8 @@
+"""Utilities for the io module."""
+
+from .pattern_resolver import PatternResolver, PatternResolverError
+
+__all__ = [
+		"PatternResolver",
+		"PatternResolverError",
+]
